Update README.md
(i) Within Ireland the language is called "Irish" not "Gaelic". However within UK (and maybe USA?) people call it "Gaelic". Using the general OSM principle of "local community usage", it's probably better to call it "Irish". ;)

(ii) Clarified how the `name` tag is used. It's basically the local name, which is English, *except* for places in the [Gaeltach](https://en.wikipedia.org/wiki/Gaeltacht), the legally defined Irish speaking areas in a few patches along the west coast. In those areas, `name` is in Irish, with `name:en` for it in English. This is not without controversy. Dingle, on the south west coast, is a large tourist destination, and in the Gaeltacht. To aid navigation, locals have been spray painting the English name on sign posts ([read more](https://en.wikipedia.org/wiki/Dingle#Name)).